+++ build_runner/lib/src/daemon/asset_server.dart
@@ -5,8 +5,9 @@
 import 'dart:async';
 import 'dart:io';
 
+import 'package:build_runner/src/daemon/daemon_builder.dart';
 import 'package:build_runner/src/server/server.dart';
-import 'package:build_runner_core/build_runner_core.dart';
+import 'package:shelf/shelf.dart';
 import 'package:shelf/shelf_io.dart' as shelf_io;
 
 class AssetServer {
@@ -19,9 +20,15 @@ class AssetServer {
   Future<void> stop() => _server.close(force: true);
 
   static Future<AssetServer> run(
-      FinalizedReader reader, String rootPackage) async {
+    BuildRunnerDaemonBuilder builder,
+    String rootPackage,
+  ) async {
     var server = await HttpServer.bind(InternetAddress.anyIPv6, 0);
-    shelf_io.serveRequests(server, AssetHandler(reader, rootPackage).handle);
+    var cascade = Cascade().add((_) async {
+      await builder.building;
+      return Response.notFound('');
+    }).add(AssetHandler(builder.reader, rootPackage).handle);
+    shelf_io.serveRequests(server, cascade.handler);
     return AssetServer._(server);
   }
 }